feat: Finish implementing lookup api for krist addresses
NOTE: PUT YOUR SCOPES BEFORE ALL THE SERVICE CALLS, THEY WILL NOT WORK OTHERWISE

+++ src/models/addresses.rs
@@ -33,7 +33,7 @@ pub struct AddressJson {
     #[serde(rename = "firstseen")]
     pub first_seen: String,
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub names: Option<i64>,
+    pub names: Option<usize>,
 }
 
 #[derive(Debug, Clone, PartialEq, PartialOrd, Serialize, Deserialize)]
